+++ mobile/android/android-components/components/browser/menu2/src/test/java/mozilla/components/browser/menu2/BrowserMenuControllerTest.kt
@@ -36,16 +36,6 @@ class BrowserMenuControllerTest {
     }
 
     @Test
-    fun `can show with custom width`() {
-        val menu = BrowserMenuController()
-        val width = 500
-        val anchor = Button(testContext)
-        val popup = menu.show(anchor, Orientation.UP, width)
-
-        assertEquals(width, popup.width)
-    }
-
-    @Test
     fun `observer is notified when submitList is called`() {
         var submitted: List<MenuCandidate>? = null
         val menu: MenuController = BrowserMenuController()